initial commit for refactor
