+++ backend/src/main/java/com/recorder/config/AIModelConfig.java
@@ -1,0 +1,6 @@
+package com.recorder.config;
+
+public class AIModelConfig {
+    public static final String MODEL_NAME = System.getenv().getOrDefault("AI_MODEL", "codellama");
+    public static final String ENDPOINT = System.getenv().getOrDefault("AI_ENDPOINT", "http://localhost:11434/api/generate");
+}
