+++ backend/src/main/java/com/recorder/config/AppConfig.java
@@ -1,0 +1,5 @@
+package com.recorder.config;
+
+public class AppConfig {
+    public static final String BASE_URL = "https://example.com";
+}
