+++ backend/src/main/java/com/recorder/driver/WebDriverFactory.java
@@ -1,0 +1,27 @@
+package com.recorder.driver;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+public class WebDriverFactory {
+    private static WebDriver driver;
+
+    public static WebDriver createDriver() {
+        if (driver == null) {
+            WebDriverManager.chromedriver().setup();
+            ChromeOptions options = new ChromeOptions();
+            options.addArguments("--disable-web-security", "--allow-running-insecure-content", "--user-data-dir=/tmp/profile");
+            driver = new ChromeDriver(options);
+        }
+        return driver;
+    }
+
+    public static void quitDriver() {
+        if (driver != null) {
+            driver.quit();
+            driver = null;
+        }
+    }
+}
