+++ backend/src/main/java/com/recorder/service/LocatorService.java
@@ -1,0 +1,49 @@
+package com.recorder.service;
+
+import com.recorder.model.BrowserAction;
+import com.recorder.model.Locator;
+import jakarta.ws.rs.core.Response;
+
+import java.util.*;
+
+public class LocatorService {
+
+    private static final List<Locator> LOCATORS = new ArrayList<>();
+
+    public static void addLocator(Locator locator) {
+        LOCATORS.add(locator);
+    }
+
+    public static void clear() {
+        LOCATORS.clear();
+    }
+
+    public static List<Locator> getRawLocators() {
+        return new ArrayList<>(LOCATORS);
+    }
+
+    public static List<Locator> getDeduplicatedLocatorsFromActions(List<BrowserAction> actions) {
+        Set<String> seen = new HashSet<>();
+        List<Locator> result = new ArrayList<>();
+
+        for (BrowserAction action : actions) {
+            String selector = action.getSelector();
+            if (selector == null || selector.isBlank()) continue;
+
+            String name = action.getTag() != null ? action.getTag() : "element";
+            String key = name + "::cssSelector::" + selector;
+
+            if (seen.add(key)) {
+                Locator locator = new Locator();
+                locator.setName(name);
+                Locator.LookupDetails details = new Locator.LookupDetails();
+                details.setFindBy("cssSelector");
+                details.setValue(selector);
+                locator.setLookupDetails(details);
+                result.add(locator);
+            }
+        }
+
+        return result;
+    }
+}
