+++ backend/src/main/java/com/recorder/util/AIClient.java
@@ -1,0 +1,24 @@
+package com.recorder.util;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.recorder.config.AIModelConfig;
+import com.recorder.model.OllamaRequest;
+
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+
+public class AIClient {
+    public static String sendToAI(OllamaRequest request) throws Exception {
+        ObjectMapper mapper = new ObjectMapper();
+        String requestBody = mapper.writeValueAsString(request);
+        HttpRequest httpRequest = HttpRequest.newBuilder()
+                .uri(URI.create(AIModelConfig.ENDPOINT))
+                .header("Content-Type", "application/json")
+                .POST(HttpRequest.BodyPublishers.ofString(requestBody))
+                .build();
+        HttpResponse<String> response = HttpClient.newHttpClient().send(httpRequest, HttpResponse.BodyHandlers.ofString());
+        return response.body();
+    }
+}
